Made both the Tool Head's Durability Multiplier and the amount the Grindstone will fix up the Head's durability each tick of grinding it configurable! Now both of these can be configured as you please.
Fixed some more hiccups with Smithing Plus compatability, should be able to properly re-use a repaired tool head and it shouldn't cause Smithing Plus to crash cause of the placeholder recipe it sends when you make a tool.

Added the default ability and patch needed to use Bone as a handle! Pretty much functions like a stick with a tiny HP bonus. But the 'bone' grid crafted tools will remain functioning like Vanilla tools for now.

Added a compat patch for Wildcraft Trees to just quickly patch that mod's storageflags patch so you can still hold sticks and bones in the offhand, and also store them in the Foraging Basket.